Fix broken test
The test was not going through the board matrix as intended,
and there is a big typo where it is referencing non existence
property and object. Author had no idea how the test was passing.
Well, he have himself to blame.

This commit also:
Construct the board matrix the bare way to ensure every slot is
not linked by reference. Yep. That thing is a huge kick in the
family jewel, ...or baby maker.